moves switch statement for showing different components to own file, related #80
